requirements.txt
tweepy
Pillow
requests
time
shutil